--- shared/modules/selectors/account.ts
@@ -1,5 +1,0 @@
-export {
-  getCurrentKeyring,
-  getHardwareWalletType,
-  isHardwareWallet,
-} from './hardware-wallet';
