+++ ui/components/multichain/account-list-item-menu/account-list-item-menu.js
@@ -6,7 +6,7 @@ import { formatChainIdToCaip } from '@metamask/bridge-controller';
 import { MetaMetricsContext } from '../../../contexts/metametrics';
 import { useI18nContext } from '../../../hooks/useI18nContext';
 import { getCurrentChainId } from '../../../../shared/modules/selectors/networks';
-import { getHardwareWalletType } from '../../../../shared/modules/selectors';
+import { getHardwareWalletType } from '../../../../shared/modules/selectors/hardware-wallet';
 import {
   getAccountTypeForKeyring,
   getPinnedAccountsList,
