+++ ui/pages/bridge/hooks/useSubmitBridgeTransaction.ts
@@ -7,7 +7,7 @@ import {
   isNonEvmChainId,
 } from '@metamask/bridge-controller';
 import type { QuoteMetadata, QuoteResponse } from '@metamask/bridge-controller';
-import { isHardwareWallet } from '../../../../shared/modules/selectors';
+import { isHardwareWallet } from '../../../../shared/modules/selectors/hardware-wallet';
 import { captureException } from '../../../../shared/lib/sentry';
 import {
   AWAITING_SIGNATURES_ROUTE,
