+++ ui/pages/confirmations/hooks/pay/useAutomaticTransactionPayToken.test.tsx
@@ -15,7 +15,7 @@ import type { SetPayTokenRequest } from './types';
 jest.mock('./useTransactionPayToken');
 jest.mock('./useTransactionPayData');
 jest.mock('./useTransactionPayAvailableTokens');
-jest.mock('../../../../../shared/modules/selectors', () => ({
+jest.mock('../../../../../shared/modules/selectors/hardware-wallet', () => ({
   getHardwareWalletType: jest.fn(() => null),
 }));
