+++ ui/pages/swaps/awaiting-swap/awaiting-swap.js
@@ -21,11 +21,13 @@ import {
 } from '../../../selectors';
 import { getHDEntropyIndex } from '../../../selectors/selectors';
 import {
-  getHardwareWalletType,
-  isHardwareWallet,
   getSmartTransactionsEnabled,
   getSmartTransactionsOptInStatusForMetrics,
 } from '../../../../shared/modules/selectors';
+import {
+  getHardwareWalletType,
+  isHardwareWallet,
+} from '../../../../shared/modules/selectors/hardware-wallet';
 
 import {
   getUsedQuote,
