+++ ui/pages/swaps/loading-swaps-quotes/loading-swaps-quotes.js
@@ -13,11 +13,13 @@ import {
 } from '../../../ducks/swaps/swaps';
 import { getHDEntropyIndex } from '../../../selectors/selectors';
 import {
-  getHardwareWalletType,
-  isHardwareWallet,
   getSmartTransactionsEnabled,
   getSmartTransactionsOptInStatusForMetrics,
 } from '../../../../shared/modules/selectors';
+import {
+  getHardwareWalletType,
+  isHardwareWallet,
+} from '../../../../shared/modules/selectors/hardware-wallet';
 import { I18nContext } from '../../../contexts/i18n';
 import { MetaMetricsContext } from '../../../contexts/metametrics';
 import Mascot from '../../../components/ui/mascot';
