+++ ui/pages/swaps/smart-transaction-status/smart-transaction-status.js
@@ -16,11 +16,13 @@ import {
 import { getCurrentChainId } from '../../../../shared/modules/selectors/networks';
 import { getRpcPrefsForCurrentProvider } from '../../../selectors';
 import {
-  getHardwareWalletType,
-  isHardwareWallet,
   getSmartTransactionsEnabled,
   getSmartTransactionsOptInStatusForMetrics,
 } from '../../../../shared/modules/selectors';
+import {
+  getHardwareWalletType,
+  isHardwareWallet,
+} from '../../../../shared/modules/selectors/hardware-wallet';
 import { CHAINID_DEFAULT_BLOCK_EXPLORER_URL_MAP } from '../../../../shared/constants/common';
 import {
   DEFAULT_ROUTE,
